Fix frequencies sticking at max with GCC 6 and 7
Leaving these variables uninitialized yielded some interesting
results... I believe it is due to GCC optimization in later
updates (as adding -DDEBUG to the makefile also fixed it though
I couldn't reproduce down to -O1). Without this patch, the min_index
will pretty much always be 21, resulting in the CPU getting pegged
at its max value and never coming down. With this patch, GCC 4.9 and
GCC 7.2 perform the same, as verified with logging.

Also, fix a boolean fail (!ret means equal to zero which is the
expected return value for that function) and a Windows carriage
return (lol).

Change-Id: I372f36cbb4416504519b51bf0d3ddff0ae584030
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>