staging: qcacld-3.0: Fix all relevant GCC warnings
This is the third part in the GCC warnings series, fixing all
warnings that come from CAF's Wi-Fi stack.

This was done with the following toolchain version:
aarch64-linaro-linux-android-gcc (Linaro GCC 7.2-2017.11~dev) 7.2.1
20171011

If a warning was a false positive, it was disabled in the folder's
Makefile. I believe all of these are correct but I am 100% open to
discussion (I am not infallible).

Change-Id: Iffe550e451bb47b0bc3ab39f005ad01ad19730fd
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>